+++ classy/__init__.py
@@ -1,22 +1,5 @@
-__version__ = 0.2
+"""Taxonomic classification of asteroid observations following Mahlke+ 2022."""
+from .preprocessing import Preprocessor
+from .classify import Classifier
 
-import importlib.util
-from . import (
-    cli,
-    classify,
-    data,
-    decision_tree,
-    defs,
-    gmm,
-    logging,
-    mcfa,
-    mixnorm,
-    plotting,
-    preprocessing,
-    spectra,
-    tools,
-)
-
-from pathlib import Path
-
-PATH_DATA = Path(importlib.util.find_spec("classy").origin).parent.parent / "data"
+__version__ = "0.2.1"