+++ CHANGELOG.md
@@ -1,2 +1,16 @@
-# 0.1 - 2022-04-
+# 0.2.1
+- Fix implementation of data/ directory
+- Simplify import scheme in package
+
+  New recommended way to interact with classy is
+
+    from classy import Preprocessor
+    from classy import Classifier
+
+- Print logging output in jupyter notebook
+
+# 0.2 - 2022-05-3
 - Initial release
+
+# 0.1 - 2021
+404